Fix STL binary writer
Need to write data in the size of float, not double, to adhere to the
STL binary file format spec

+++ src/IO/StlBinaryWriter.cs
@@ -26,14 +26,14 @@ namespace ScaffoldingGenerator.IO
                 foreach (Facet facet in facets)
                 {
                     Vector3D normal = facet.Normal;
-                    writer.Write(normal.X);
-                    writer.Write(normal.Y);
-                    writer.Write(normal.Z);
+                    writer.Write((float)normal.X);
+                    writer.Write((float)normal.Y);
+                    writer.Write((float)normal.Z);
                     foreach (Point3D vertex in facet.Vertices)
                     {
-                        writer.Write(vertex.X);
-                        writer.Write(vertex.Y);
-                        writer.Write(vertex.Z);
+                        writer.Write((float)vertex.X);
+                        writer.Write((float)vertex.Y);
+                        writer.Write((float)vertex.Z);
                     }
                     writer.Write(new byte[2]);
                 }
